MDL-41652 backup: Dates should be rolled as expected
Make sure activity dates are rolled as expected. The issue was the default value for coursestartdate
was set incorrectly in certain cases, this patch fixes that. Also this patch makes sure that the
coursestartdate is set to false for all cases except when both overwrite conf and startdate are set